Feat: 매거진 products 모델 및 로직

+++ products/views.py
@@ -8,27 +8,26 @@ from .serializers import DrfProductSerializer
 from rest_framework.parsers import MultiPartParser, FormParser
 import logging
 
-logger = logging.getLogger('django')
-
 
 class CreateProductView(APIView):
     permission_classes = [IsAuthenticated]
     parser_classes = (MultiPartParser, FormParser)
 
     def post(self, request):
+        # 사용자가 로그인했는지 확인
         if not request.user.is_authenticated:
             return Response({"detail": "Authentication required"}, status=status.HTTP_401_UNAUTHORIZED)
 
-        logger.debug("Request data: %s", request.data)
+        # 제품 작성 권한이 있는지 확인
+        if not request.user.is_product_writer:
+            return Response({"detail": "You do not have permission to create a product."}, status=status.HTTP_403_FORBIDDEN)
         
         serializer = DrfProductSerializer(data=request.data, context={'request': request})
         if serializer.is_valid():
-            serializer.save(owner=request.user)
+            serializer.save(writer=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
-        else:
-            logger.debug("Serializer errors: %s", serializer.errors)
+        
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
 
 
 class UpdateProductView(APIView):
@@ -36,7 +35,11 @@ class UpdateProductView(APIView):
     parser_classes = (MultiPartParser, FormParser)
 
     def put(self, request, id):
-        product = get_object_or_404(DrfProduct, id=id, owner=request.user) 
+        
+        if not request.user.is_product_writer:
+            return Response({"detail": "You do not have permission to update a product."}, status=status.HTTP_403_FORBIDDEN)
+
+        product = get_object_or_404(DrfProduct, productId=id, writer=request.user)
         serializer = DrfProductSerializer(product, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
@@ -49,7 +52,7 @@ class RetrieveProductView(APIView):
     permission_classes = [IsAuthenticated]
 
     def get(self, request, id):
-        product = get_object_or_404(DrfProduct, id=id, owner=request.user)  # 유저가 소유한 제품인지 확인
+        product = get_object_or_404(DrfProduct, productId=id)  # writer 조건 제거
         serializer = DrfProductSerializer(product, context={'request': request})
         return Response(serializer.data, status=status.HTTP_200_OK)
 
@@ -59,20 +62,25 @@ class DeleteProductView(APIView):
     permission_classes = [IsAuthenticated]
 
     def delete(self, request, id):
-        product = get_object_or_404(DrfProduct, id=id, owner=request.user)  # 유저가 소유한 제품인지 확인
+        # 제품 작성 권한이 있는지 확인
+        if not request.user.is_product_writer:
+            return Response({"detail": "You do not have permission to delete a product."}, status=status.HTTP_403_FORBIDDEN)
+
+        product = get_object_or_404(DrfProduct, productId=id, writer=request.user)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
 
 # 사용자가 소유한 모든 제품 리스트 조회
 class ListProductsView(APIView):
     permission_classes = [IsAuthenticated]
 
     def get(self, request):
-        # 요청한 사용자의 모든 제품을 필터링합니다.
-        products = DrfProduct.objects.filter(owner=request.user)
-        
+        # 모든 사용자의 제품을 조회
+        products = DrfProduct.objects.all()
+
         if not products.exists():
-            return Response({'detail': 'No products found for this user.'}, status=status.HTTP_404_NOT_FOUND)
+            return Response({'detail': 'No products found.'}, status=status.HTTP_404_NOT_FOUND)
 
         serializer = DrfProductSerializer(products, many=True, context={'request': request})
         return Response(serializer.data, status=status.HTTP_200_OK)
